% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_url.R
\name{browse_url}
\alias{browse_url}
\alias{browse_url.cs_species}
\alias{browse_url,cs_species-method}
\title{Load species page in browser}
\usage{
browse_url(x, ...)

\method{browse_url}{cs_species}(x, maxsum = 10, ...)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object.}

\item{...}{Further arguments passed to \code{\link[=browseURL]{browseURL()}}.}

\item{maxsum}{An integer value indicating the maximum number of URLs to be
opened at once. This parameter avoids opening too much tabs in your
browser.}
}
\description{
Load the respective page of a species or taxon into the default browser.
This is a wrapper of \code{\link[=browseURL]{browseURL()}}.
}
