% multi-purpose template
\name{conosur-package}
\docType{package}
\alias{conosur-package}
\alias{conosur}

\title{Retrieve Data from \dQuote{Flora del Conosur}.}
\description{
This package contains a set of functions to retrieve data from
\dQuote{Flora del Conosur}, which is available at
\url{http://www.darwin.edu.ar/Proyectos/FloraArgentina/fa.htm}.
}
\author{
Miguel Alvarez, \email{malvarez@uni-bonn.de}
}
