% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{cs_species-class}
\alias{cs_species-class}
\title{List of species}
\description{
An S3 class for species lists retrieved by \code{\link[=conosur_species]{conosur_species()}}, which
inherits properties of data frames.
}
