% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_author.R
\name{get_author}
\alias{get_author}
\title{Appends authorships to existing lists of species}
\usage{
get_author(x)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object of one row.}
}
\description{
Author names need to be retrieved from the specific pages of species in an
additional step. This function is called internally.
}
\keyword{internal}
