% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\alias{get_info.cs_species}
\alias{get_info,cs_species-method}
\title{Add authorship and attributes to a species list}
\usage{
get_info(x, ...)

\method{get_info}{cs_species}(x, subset, progress = TRUE, ...)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object.}

\item{...}{Further arguments passed among methods.}

\item{subset}{An expression selecting a subset of rows for retrieving the
information. This can be useful if a previous run had failed in some
species.}

\item{progress}{A logical value indicating whether a progress bar should be
displayed or not.}
}
\value{
A \linkS4class{cs_species} object with additional columns named \strong{AuthorName}
(species name authorship), \strong{status} (native or non-native), \strong{life_form},
\strong{elevation_min} and \strong{elevation_max}.
}
\description{
Author names and species attributes of species lists downloaded by
\code{\link[=conosur_species]{conosur_species()}} need to be requested in specific pages.
}
\examples{
## List for letter K
splist <- conosur_species(letter = "K")[1:5, ]
splist <- get_info(splist)

}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}
}
