% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonyms.R
\name{get_synonyms}
\alias{get_synonyms}
\alias{get_synonyms.cs_species}
\alias{get_synonyms,cs_species-method}
\title{Get synonyms for retrieved species lists}
\usage{
get_synonyms(x, ...)

\method{get_synonyms}{cs_species}(x, progress = TRUE, ...)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object.}

\item{...}{Further arguments passed among methods.}

\item{progress}{A logical value indicating whether a progress bar should be
displayed or not.}
}
\value{
The input \linkS4class{cs_species} object with appended rows for synonyms.
An additional column \strong{AcceptedName} is added indicating accepted names with
value \code{'TRUE'} and synonyms with value \code{'FALSE'}.
}
\description{
Get synonyms from species lists retrieved by \code{\link[=conosur_species]{conosur_species()}}.
}
\examples{
## List for letter K
splist <- conosur_species(letter = "K")[1:5, ]
splist <- get_synonyms(splist)
head(splist)

}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}
}
