% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_table}
\alias{get_table}
\title{Appends attributes to an existing lists of species}
\usage{
get_table(x)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object of one row.}
}
\description{
Author names need to be retrieved from the specific pages of species in an
additional step. This function is called internally.
}
\keyword{internal}
