% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conosur_species.R
\name{get_url}
\alias{get_url}
\title{Get url for the species list}
\usage{
get_url(home, letter)
}
\arguments{
\item{home}{Character value with the URL for the main site.}

\item{letter}{Character vector with the requested initials.}
}
\description{
In the Flora del Conosur species are listed in single pages by letters
(initials). A first step to retrieve the list is getting the respective
URLs.
}
\keyword{internal}
