% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synonyms.R
\name{get_url_syno}
\alias{get_url_syno}
\title{Getting synonyms for retrieved taxonomic lists}
\usage{
get_url_syno(home, x)
}
\arguments{
\item{home}{Character value with the home url.}

\item{x}{A \linkS4class{cs_species} object of one row.}
}
\description{
Lists of synonyms are scrapped from the species list.
}
\keyword{internal}
