% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_author.R
\name{load_author}
\alias{load_author}
\alias{load_author.cs_species}
\alias{load_author,cs_species-method}
\title{Add authorship to a species list}
\usage{
load_author(x, ...)

\method{load_author}{cs_species}(x, progress = TRUE, ...)
}
\arguments{
\item{x}{A \linkS4class{cs_species} object.}

\item{...}{Further arguments passed among methods.}

\item{progress}{A logical value indicating whether a progress bar should be
displayed or not.}
}
\value{
A \linkS4class{cs_species} object with an additional column named \code{'AuthorName'}.
}
\description{
Add author names of species lists loaded by \code{\link[=conosur_species]{conosur_species()}}.
}
\examples{
## List for letter A
splist <- conosur_species(letter = "K")[1:5, ]
splist <- load_author(splist)

}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}
}
