\name{tropicos_specimens}
\alias{tropicos_specimens}

\title{Query for specimens in \dQuote{Tropicos}.}
\description{
Documentation under construction
}
\usage{
tropicos_specimens(Home="http://www.tropicos.org/SpecimenSearch.aspx",
		collector, number, barcode, accession, specimen, mode="collector",
		encoding="UTF-8")
}
\arguments{
\item{Home}{Site for search (default).}
\item{collector}{Name of senior collector.}
\item{number}{Collection number by respective senior collector.}
\item{barcode}{Barcode ID.}
\item{accession}{Number of accession.}
\item{specimen}{Specimen ID.}
\item{mode}{Entry used to retrieve specimen.}
\item{encoding}{Character encoding.}
}
\details{
Documentation under construction
}
\value{
Documentation under construction
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}.
}
\examples{
library(conosur)

## Documentation under construction
}

