% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conosur-package.R
\docType{package}
\name{conosur-package}
\alias{conosur-package}
\alias{conosur}
\title{Retrieve Data from the Flora del Conosur.}
\description{
A set of functions to retrieve data from
\href{http://www.darwin.edu.ar/Proyectos/FloraArgentina/fa.htm}{Flora del Conosur}.
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}
}
