% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conosur_species.R
\name{conosur_species}
\alias{conosur_species}
\title{Main species list from the Flora del Conosur}
\usage{
conosur_species(letter = LETTERS, progress = TRUE)
}
\arguments{
\item{letter}{Character vector indicating the initials of requested names.}

\item{progress}{Logical value indicating whether a progress bar should be
initiated or not. The progress bar will be called by \code{\link[=tkProgressBar]{tkProgressBar()}}.
If length of argument \code{'letter'} is 1, no progress bar will be called.}
}
\value{
An object of class \linkS4class{cs_species}, which is in fact a data frame
containing following variables:
\describe{
\item{TaxonUsageID}{ID of the species names in Flora del Conosur.}
\item{TaxonConceptID}{ID of the taxon concept in Flora del Conosur
(the same as TaxonUsageID).}
\item{TaxonName}{Full name of the taxon (excluding author name).}
\item{genus}{Genus name.}
\item{species}{Species name.}
\item{subspecies}{Subspecies name for subspecies.}
\item{variety}{Variety name for varieties.}
\item{form}{Form name for forms.}
\item{url}{Link to the specific taxon page.}
}
}
\description{
Retrieve species list from the alphabetic overview in
\href{http://www.darwin.edu.ar/Proyectos/FloraArgentina/Especies.asp}{Flora del Conosur}.
The species list is organized in the page by initials of the genera and can
be downloaded letter-wise.
}
\examples{
## List for letter A
splist <- conosur_species(letter = "Z")
head(splist)
tail(splist)

## This function is also working with words
splist <- conosur_species(letter = "kilo")
head(splist)
tail(splist)

}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}
}
